+++ lib/sfml/c/graphics.rb
@@ -97,6 +97,8 @@ module SFML
       attach_function :sfRenderWindow_getSize,                [:render_window_t], System::Vector2u.by_value
       attach_function :sfRenderWindow_setSize,                [:render_window_t, System::Vector2u.by_value], :void
       attach_function :sfRenderWindow_setIcon,                [:render_window_t, System::Vector2u.by_value, :pointer], :void
+      attach_function :sfRenderWindow_setMinimumSize,         [:render_window_t, :pointer], :void
+      attach_function :sfRenderWindow_setMaximumSize,         [:render_window_t, :pointer], :void
 
       typedef :pointer, :texture_t
       typedef :pointer, :render_texture_t